third_party/WebKit/LayoutTests/virtual/threaded/http/tests/worklet/webexposed/global-interface-listing-paint-worklet-expected.txt

CONSOLE MESSAGE: line 10: This test logs exposed APIs once from each PaintWorkletGlobalScope
CONSOLE MESSAGE: line 151: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method size
CONSOLE MESSAGE: line 151: interface CSSImageValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSKeywordValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     getter value
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter value
CONSOLE MESSAGE: line 151: interface CSSMathInvert : CSSMathValue
CONSOLE MESSAGE: line 151:     getter value
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathMax : CSSMathValue
CONSOLE MESSAGE: line 151:     getter values
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathMin : CSSMathValue
CONSOLE MESSAGE: line 151:     getter values
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathNegate : CSSMathValue
CONSOLE MESSAGE: line 151:     getter value
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathProduct : CSSMathValue
CONSOLE MESSAGE: line 151:     getter values
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathSum : CSSMathValue
CONSOLE MESSAGE: line 151:     getter values
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathValue : CSSNumericValue
CONSOLE MESSAGE: line 151:     getter operator
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMatrixComponent : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter matrix
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter matrix
CONSOLE MESSAGE: line 151: interface CSSNumericArray
CONSOLE MESSAGE: line 151:     getter length
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method entries
CONSOLE MESSAGE: line 151:     method forEach
CONSOLE MESSAGE: line 151:     method keys
CONSOLE MESSAGE: line 151:     method values
CONSOLE MESSAGE: line 151: interface CSSNumericValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     method add
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method div
CONSOLE MESSAGE: line 151:     method equals
CONSOLE MESSAGE: line 151:     method max
CONSOLE MESSAGE: line 151:     method min
CONSOLE MESSAGE: line 151:     method mul
CONSOLE MESSAGE: line 151:     method sub
CONSOLE MESSAGE: line 151:     method to
CONSOLE MESSAGE: line 151:     method toSum
CONSOLE MESSAGE: line 151:     method type
CONSOLE MESSAGE: line 151: interface CSSPerspective : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter length
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter length
CONSOLE MESSAGE: line 151: interface CSSPositionValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     getter x
CONSOLE MESSAGE: line 151:     getter y
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter x
CONSOLE MESSAGE: line 151:     setter y
CONSOLE MESSAGE: line 151: interface CSSRotate : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter angle
CONSOLE MESSAGE: line 151:     getter x
CONSOLE MESSAGE: line 151:     getter y
CONSOLE MESSAGE: line 151:     getter z
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter angle
CONSOLE MESSAGE: line 151:     setter x
CONSOLE MESSAGE: line 151:     setter y
CONSOLE MESSAGE: line 151:     setter z
CONSOLE MESSAGE: line 151: interface CSSScale : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter x
CONSOLE MESSAGE: line 151:     getter y
CONSOLE MESSAGE: line 151:     getter z
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter x
CONSOLE MESSAGE: line 151:     setter y
CONSOLE MESSAGE: line 151:     setter z
CONSOLE MESSAGE: line 151: interface CSSSkew : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter ax
CONSOLE MESSAGE: line 151:     getter ay
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter ax
CONSOLE MESSAGE: line 151:     setter ay
CONSOLE MESSAGE: line 151: interface CSSStyleValue
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method toString
CONSOLE MESSAGE: line 151: interface CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter is2D
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method toMatrix
CONSOLE MESSAGE: line 151:     method toString
CONSOLE MESSAGE: line 151:     setter is2D
CONSOLE MESSAGE: line 151: interface CSSTransformValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     getter is2D
CONSOLE MESSAGE: line 151:     getter length
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method entries
CONSOLE MESSAGE: line 151:     method forEach
CONSOLE MESSAGE: line 151:     method keys
CONSOLE MESSAGE: line 151:     method toMatrix
CONSOLE MESSAGE: line 151:     method values
CONSOLE MESSAGE: line 151: interface CSSTranslate : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter x
CONSOLE MESSAGE: line 151:     getter y
CONSOLE MESSAGE: line 151:     getter z
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter x
CONSOLE MESSAGE: line 151:     setter y
CONSOLE MESSAGE: line 151:     setter z
CONSOLE MESSAGE: line 151: interface CSSUnitValue : CSSNumericValue
CONSOLE MESSAGE: line 151:     getter unit
CONSOLE MESSAGE: line 151:     getter value
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter value
CONSOLE MESSAGE: line 151: interface CSSUnparsedValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     getter length
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method entries
CONSOLE MESSAGE: line 151:     method forEach
CONSOLE MESSAGE: line 151:     method keys
CONSOLE MESSAGE: line 151:     method values
CONSOLE MESSAGE: line 151: interface CSSVariableReferenceValue
CONSOLE MESSAGE: line 151:     getter fallback
CONSOLE MESSAGE: line 151:     getter variable
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter variable
CONSOLE MESSAGE: line 151: interface CountQueuingStrategy
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method size
CONSOLE MESSAGE: line 151: interface PaintRenderingContext2D
CONSOLE MESSAGE: line 151:     getter fillStyle
CONSOLE MESSAGE: line 151:     getter filter
CONSOLE MESSAGE: line 151:     getter globalAlpha
CONSOLE MESSAGE: line 151:     getter globalCompositeOperation
CONSOLE MESSAGE: line 151:     getter imageSmoothingEnabled
CONSOLE MESSAGE: line 151:     getter imageSmoothingQuality
CONSOLE MESSAGE: line 151:     getter lineCap
CONSOLE MESSAGE: line 151:     getter lineDashOffset
CONSOLE MESSAGE: line 151:     getter lineJoin
CONSOLE MESSAGE: line 151:     getter lineWidth
CONSOLE MESSAGE: line 151:     getter miterLimit
CONSOLE MESSAGE: line 151:     getter shadowBlur
CONSOLE MESSAGE: line 151:     getter shadowColor
CONSOLE MESSAGE: line 151:     getter shadowOffsetX
CONSOLE MESSAGE: line 151:     getter shadowOffsetY
CONSOLE MESSAGE: line 151:     getter strokeStyle
CONSOLE MESSAGE: line 151:     method arc
CONSOLE MESSAGE: line 151:     method arcTo
CONSOLE MESSAGE: line 151:     method beginPath
CONSOLE MESSAGE: line 151:     method bezierCurveTo
CONSOLE MESSAGE: line 151:     method clearRect
CONSOLE MESSAGE: line 151:     method clip
CONSOLE MESSAGE: line 151:     method closePath
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method createLinearGradient
CONSOLE MESSAGE: line 151:     method createPattern
CONSOLE MESSAGE: line 151:     method createRadialGradient
CONSOLE MESSAGE: line 151:     method drawImage
CONSOLE MESSAGE: line 151:     method ellipse
CONSOLE MESSAGE: line 151:     method fill
CONSOLE MESSAGE: line 151:     method fillRect
CONSOLE MESSAGE: line 151:     method getLineDash
CONSOLE MESSAGE: line 151:     method getTransform
CONSOLE MESSAGE: line 151:     method isPointInPath
CONSOLE MESSAGE: line 151:     method isPointInStroke
CONSOLE MESSAGE: line 151:     method lineTo
CONSOLE MESSAGE: line 151:     method moveTo
CONSOLE MESSAGE: line 151:     method quadraticCurveTo
CONSOLE MESSAGE: line 151:     method rect
CONSOLE MESSAGE: line 151:     method resetTransform
CONSOLE MESSAGE: line 151:     method restore
CONSOLE MESSAGE: line 151:     method rotate
CONSOLE MESSAGE: line 151:     method save
CONSOLE MESSAGE: line 151:     method scale
CONSOLE MESSAGE: line 151:     method setLineDash
CONSOLE MESSAGE: line 151:     method setTransform
CONSOLE MESSAGE: line 151:     method stroke
CONSOLE MESSAGE: line 151:     method strokeRect
CONSOLE MESSAGE: line 151:     method transform
CONSOLE MESSAGE: line 151:     method translate
CONSOLE MESSAGE: line 151:     setter fillStyle
CONSOLE MESSAGE: line 151:     setter filter
CONSOLE MESSAGE: line 151:     setter globalAlpha
CONSOLE MESSAGE: line 151:     setter globalCompositeOperation
CONSOLE MESSAGE: line 151:     setter imageSmoothingEnabled
CONSOLE MESSAGE: line 151:     setter imageSmoothingQuality
CONSOLE MESSAGE: line 151:     setter lineCap
CONSOLE MESSAGE: line 151:     setter lineDashOffset
CONSOLE MESSAGE: line 151:     setter lineJoin
CONSOLE MESSAGE: line 151:     setter lineWidth
CONSOLE MESSAGE: line 151:     setter miterLimit
CONSOLE MESSAGE: line 151:     setter shadowBlur
CONSOLE MESSAGE: line 151:     setter shadowColor
CONSOLE MESSAGE: line 151:     setter shadowOffsetX
CONSOLE MESSAGE: line 151:     setter shadowOffsetY
CONSOLE MESSAGE: line 151:     setter strokeStyle
CONSOLE MESSAGE: line 151: interface PaintSize
CONSOLE MESSAGE: line 151:     getter height
CONSOLE MESSAGE: line 151:     getter width
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface PaintWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface Path2D
CONSOLE MESSAGE: line 151:     method addPath
CONSOLE MESSAGE: line 151:     method arc
CONSOLE MESSAGE: line 151:     method arcTo
CONSOLE MESSAGE: line 151:     method bezierCurveTo
CONSOLE MESSAGE: line 151:     method closePath
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method ellipse
CONSOLE MESSAGE: line 151:     method lineTo
CONSOLE MESSAGE: line 151:     method moveTo
CONSOLE MESSAGE: line 151:     method quadraticCurveTo
CONSOLE MESSAGE: line 151:     method rect
CONSOLE MESSAGE: line 151: interface ReadableStream
CONSOLE MESSAGE: line 151:     getter locked
CONSOLE MESSAGE: line 151:     method cancel
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method getReader
CONSOLE MESSAGE: line 151:     method pipeThrough
CONSOLE MESSAGE: line 151:     method pipeTo
CONSOLE MESSAGE: line 151:     method tee
CONSOLE MESSAGE: line 151: interface StylePropertyMapReadOnly
CONSOLE MESSAGE: line 151:     getter size
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method entries
CONSOLE MESSAGE: line 151:     method forEach
CONSOLE MESSAGE: line 151:     method get
CONSOLE MESSAGE: line 151:     method getAll
CONSOLE MESSAGE: line 151:     method has
CONSOLE MESSAGE: line 151:     method keys
CONSOLE MESSAGE: line 151:     method values
CONSOLE MESSAGE: line 151: interface TransformStream
CONSOLE MESSAGE: line 151:     getter readable
CONSOLE MESSAGE: line 151:     getter writable
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface WorkletGlobalScope
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface WritableStream
CONSOLE MESSAGE: line 151:     getter locked
CONSOLE MESSAGE: line 151:     method abort
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method getWriter
CONSOLE MESSAGE: line 151: global object
CONSOLE MESSAGE: line 151:     attribute console
CONSOLE MESSAGE: line 151:     getter devicePixelRatio
CONSOLE MESSAGE: line 151:     method gc
CONSOLE MESSAGE: line 151:     method registerPaint
CONSOLE MESSAGE: line 151: interface ByteLengthQueuingStrategy
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method size
CONSOLE MESSAGE: line 151: interface CSSImageValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSKeywordValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     getter value
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter value
CONSOLE MESSAGE: line 151: interface CSSMathInvert : CSSMathValue
CONSOLE MESSAGE: line 151:     getter value
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathMax : CSSMathValue
CONSOLE MESSAGE: line 151:     getter values
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathMin : CSSMathValue
CONSOLE MESSAGE: line 151:     getter values
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathNegate : CSSMathValue
CONSOLE MESSAGE: line 151:     getter value
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathProduct : CSSMathValue
CONSOLE MESSAGE: line 151:     getter values
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathSum : CSSMathValue
CONSOLE MESSAGE: line 151:     getter values
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMathValue : CSSNumericValue
CONSOLE MESSAGE: line 151:     getter operator
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface CSSMatrixComponent : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter matrix
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter matrix
CONSOLE MESSAGE: line 151: interface CSSNumericArray
CONSOLE MESSAGE: line 151:     getter length
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method entries
CONSOLE MESSAGE: line 151:     method forEach
CONSOLE MESSAGE: line 151:     method keys
CONSOLE MESSAGE: line 151:     method values
CONSOLE MESSAGE: line 151: interface CSSNumericValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     method add
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method div
CONSOLE MESSAGE: line 151:     method equals
CONSOLE MESSAGE: line 151:     method max
CONSOLE MESSAGE: line 151:     method min
CONSOLE MESSAGE: line 151:     method mul
CONSOLE MESSAGE: line 151:     method sub
CONSOLE MESSAGE: line 151:     method to
CONSOLE MESSAGE: line 151:     method toSum
CONSOLE MESSAGE: line 151:     method type
CONSOLE MESSAGE: line 151: interface CSSPerspective : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter length
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter length
CONSOLE MESSAGE: line 151: interface CSSPositionValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     getter x
CONSOLE MESSAGE: line 151:     getter y
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter x
CONSOLE MESSAGE: line 151:     setter y
CONSOLE MESSAGE: line 151: interface CSSRotate : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter angle
CONSOLE MESSAGE: line 151:     getter x
CONSOLE MESSAGE: line 151:     getter y
CONSOLE MESSAGE: line 151:     getter z
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter angle
CONSOLE MESSAGE: line 151:     setter x
CONSOLE MESSAGE: line 151:     setter y
CONSOLE MESSAGE: line 151:     setter z
CONSOLE MESSAGE: line 151: interface CSSScale : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter x
CONSOLE MESSAGE: line 151:     getter y
CONSOLE MESSAGE: line 151:     getter z
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter x
CONSOLE MESSAGE: line 151:     setter y
CONSOLE MESSAGE: line 151:     setter z
CONSOLE MESSAGE: line 151: interface CSSSkew : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter ax
CONSOLE MESSAGE: line 151:     getter ay
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter ax
CONSOLE MESSAGE: line 151:     setter ay
CONSOLE MESSAGE: line 151: interface CSSStyleValue
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method toString
CONSOLE MESSAGE: line 151: interface CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter is2D
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method toMatrix
CONSOLE MESSAGE: line 151:     method toString
CONSOLE MESSAGE: line 151:     setter is2D
CONSOLE MESSAGE: line 151: interface CSSTransformValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     getter is2D
CONSOLE MESSAGE: line 151:     getter length
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method entries
CONSOLE MESSAGE: line 151:     method forEach
CONSOLE MESSAGE: line 151:     method keys
CONSOLE MESSAGE: line 151:     method toMatrix
CONSOLE MESSAGE: line 151:     method values
CONSOLE MESSAGE: line 151: interface CSSTranslate : CSSTransformComponent
CONSOLE MESSAGE: line 151:     getter x
CONSOLE MESSAGE: line 151:     getter y
CONSOLE MESSAGE: line 151:     getter z
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter x
CONSOLE MESSAGE: line 151:     setter y
CONSOLE MESSAGE: line 151:     setter z
CONSOLE MESSAGE: line 151: interface CSSUnitValue : CSSNumericValue
CONSOLE MESSAGE: line 151:     getter unit
CONSOLE MESSAGE: line 151:     getter value
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter value
CONSOLE MESSAGE: line 151: interface CSSUnparsedValue : CSSStyleValue
CONSOLE MESSAGE: line 151:     getter length
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method entries
CONSOLE MESSAGE: line 151:     method forEach
CONSOLE MESSAGE: line 151:     method keys
CONSOLE MESSAGE: line 151:     method values
CONSOLE MESSAGE: line 151: interface CSSVariableReferenceValue
CONSOLE MESSAGE: line 151:     getter fallback
CONSOLE MESSAGE: line 151:     getter variable
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     setter variable
CONSOLE MESSAGE: line 151: interface CountQueuingStrategy
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method size
CONSOLE MESSAGE: line 151: interface PaintRenderingContext2D
CONSOLE MESSAGE: line 151:     getter fillStyle
CONSOLE MESSAGE: line 151:     getter filter
CONSOLE MESSAGE: line 151:     getter globalAlpha
CONSOLE MESSAGE: line 151:     getter globalCompositeOperation
CONSOLE MESSAGE: line 151:     getter imageSmoothingEnabled
CONSOLE MESSAGE: line 151:     getter imageSmoothingQuality
CONSOLE MESSAGE: line 151:     getter lineCap
CONSOLE MESSAGE: line 151:     getter lineDashOffset
CONSOLE MESSAGE: line 151:     getter lineJoin
CONSOLE MESSAGE: line 151:     getter lineWidth
CONSOLE MESSAGE: line 151:     getter miterLimit
CONSOLE MESSAGE: line 151:     getter shadowBlur
CONSOLE MESSAGE: line 151:     getter shadowColor
CONSOLE MESSAGE: line 151:     getter shadowOffsetX
CONSOLE MESSAGE: line 151:     getter shadowOffsetY
CONSOLE MESSAGE: line 151:     getter strokeStyle
CONSOLE MESSAGE: line 151:     method arc
CONSOLE MESSAGE: line 151:     method arcTo
CONSOLE MESSAGE: line 151:     method beginPath
CONSOLE MESSAGE: line 151:     method bezierCurveTo
CONSOLE MESSAGE: line 151:     method clearRect
CONSOLE MESSAGE: line 151:     method clip
CONSOLE MESSAGE: line 151:     method closePath
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method createLinearGradient
CONSOLE MESSAGE: line 151:     method createPattern
CONSOLE MESSAGE: line 151:     method createRadialGradient
CONSOLE MESSAGE: line 151:     method drawImage
CONSOLE MESSAGE: line 151:     method ellipse
CONSOLE MESSAGE: line 151:     method fill
CONSOLE MESSAGE: line 151:     method fillRect
CONSOLE MESSAGE: line 151:     method getLineDash
CONSOLE MESSAGE: line 151:     method getTransform
CONSOLE MESSAGE: line 151:     method isPointInPath
CONSOLE MESSAGE: line 151:     method isPointInStroke
CONSOLE MESSAGE: line 151:     method lineTo
CONSOLE MESSAGE: line 151:     method moveTo
CONSOLE MESSAGE: line 151:     method quadraticCurveTo
CONSOLE MESSAGE: line 151:     method rect
CONSOLE MESSAGE: line 151:     method resetTransform
CONSOLE MESSAGE: line 151:     method restore
CONSOLE MESSAGE: line 151:     method rotate
CONSOLE MESSAGE: line 151:     method save
CONSOLE MESSAGE: line 151:     method scale
CONSOLE MESSAGE: line 151:     method setLineDash
CONSOLE MESSAGE: line 151:     method setTransform
CONSOLE MESSAGE: line 151:     method stroke
CONSOLE MESSAGE: line 151:     method strokeRect
CONSOLE MESSAGE: line 151:     method transform
CONSOLE MESSAGE: line 151:     method translate
CONSOLE MESSAGE: line 151:     setter fillStyle
CONSOLE MESSAGE: line 151:     setter filter
CONSOLE MESSAGE: line 151:     setter globalAlpha
CONSOLE MESSAGE: line 151:     setter globalCompositeOperation
CONSOLE MESSAGE: line 151:     setter imageSmoothingEnabled
CONSOLE MESSAGE: line 151:     setter imageSmoothingQuality
CONSOLE MESSAGE: line 151:     setter lineCap
CONSOLE MESSAGE: line 151:     setter lineDashOffset
CONSOLE MESSAGE: line 151:     setter lineJoin
CONSOLE MESSAGE: line 151:     setter lineWidth
CONSOLE MESSAGE: line 151:     setter miterLimit
CONSOLE MESSAGE: line 151:     setter shadowBlur
CONSOLE MESSAGE: line 151:     setter shadowColor
CONSOLE MESSAGE: line 151:     setter shadowOffsetX
CONSOLE MESSAGE: line 151:     setter shadowOffsetY
CONSOLE MESSAGE: line 151:     setter strokeStyle
CONSOLE MESSAGE: line 151: interface PaintSize
CONSOLE MESSAGE: line 151:     getter height
CONSOLE MESSAGE: line 151:     getter width
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface PaintWorkletGlobalScope : WorkletGlobalScope
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface Path2D
CONSOLE MESSAGE: line 151:     method addPath
CONSOLE MESSAGE: line 151:     method arc
CONSOLE MESSAGE: line 151:     method arcTo
CONSOLE MESSAGE: line 151:     method bezierCurveTo
CONSOLE MESSAGE: line 151:     method closePath
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method ellipse
CONSOLE MESSAGE: line 151:     method lineTo
CONSOLE MESSAGE: line 151:     method moveTo
CONSOLE MESSAGE: line 151:     method quadraticCurveTo
CONSOLE MESSAGE: line 151:     method rect
CONSOLE MESSAGE: line 151: interface ReadableStream
CONSOLE MESSAGE: line 151:     getter locked
CONSOLE MESSAGE: line 151:     method cancel
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method getReader
CONSOLE MESSAGE: line 151:     method pipeThrough
CONSOLE MESSAGE: line 151:     method pipeTo
CONSOLE MESSAGE: line 151:     method tee
CONSOLE MESSAGE: line 151: interface StylePropertyMapReadOnly
CONSOLE MESSAGE: line 151:     getter size
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method entries
CONSOLE MESSAGE: line 151:     method forEach
CONSOLE MESSAGE: line 151:     method get
CONSOLE MESSAGE: line 151:     method getAll
CONSOLE MESSAGE: line 151:     method has
CONSOLE MESSAGE: line 151:     method keys
CONSOLE MESSAGE: line 151:     method values
CONSOLE MESSAGE: line 151: interface TransformStream
CONSOLE MESSAGE: line 151:     getter readable
CONSOLE MESSAGE: line 151:     getter writable
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface WorkletGlobalScope
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151: interface WritableStream
CONSOLE MESSAGE: line 151:     getter locked
CONSOLE MESSAGE: line 151:     method abort
CONSOLE MESSAGE: line 151:     method constructor
CONSOLE MESSAGE: line 151:     method getWriter
CONSOLE MESSAGE: line 151: global object
CONSOLE MESSAGE: line 151:     attribute console
CONSOLE MESSAGE: line 151:     getter devicePixelRatio
CONSOLE MESSAGE: line 151:     method gc
CONSOLE MESSAGE: line 151:     method registerPaint
